upload a demonstrate about shared_ptr conversions
This file is to demonstrate the usage of std::dynamic_pointer_cast.
A base class and a deriving class are defined, and in the main function, a shared_ptr pointing to the deriving class is allocated and then managed by a pointer of its base class. Finally, we executed  std::dynamic_pointer_cast to check what will happen.